+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/controlPropertiesEditor/ArrayValueEditor.java
@@ -14,6 +14,9 @@ import com.kaylerrenslow.armaDialogCreator.control.sv.SVStringArray;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.inputfield.ArmaStringChecker;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.inputfield.InputField;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.inputfield.StringChecker;
+import com.kaylerrenslow.armaDialogCreator.util.ReadOnlyValueObserver;
+import com.kaylerrenslow.armaDialogCreator.util.ValueListener;
+import com.kaylerrenslow.armaDialogCreator.util.ValueObserver;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.scene.Node;
@@ -21,6 +24,7 @@ import javafx.scene.control.Button;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 
@@ -28,9 +32,15 @@ import java.util.ArrayList;
  Created by Kayler on 07/13/2016.
  */
 public class ArrayValueEditor implements ValueEditor<SVStringArray> {
-	
+
+	private final ValueListener editorValueUpdateListener = new ValueListener() {
+		@Override
+		public void valueUpdated(@NotNull ValueObserver observer, Object oldValue, Object newValue) {
+			valueObserver.updateValue(createValue());
+		}
+	};
 	protected ArrayList<InputField<ArmaStringChecker, String>> editors = new ArrayList<>();
-	
+
 	protected final Button btnDecreaseSize = new Button("-");
 	protected final Button btnIncreaseSize = new Button("+");
 	private final double gap = 5;
@@ -38,14 +48,16 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 	protected final FlowPane editorsPane = new FlowPane(gap, gap);
 	private final HBox masterPane;
 	private final InputField<StringChecker, String> overrideField = new InputField<>(new StringChecker());
-	
+
+	private final ValueObserver<SVStringArray> valueObserver = new ValueObserver<>(null);
+
 	public ArrayValueEditor(int numInitialFields) {
 		masterPane = new HBox(5, editorsPane);
 		editorsPane.minWidth(0d);
 		editorsPane.prefWidth(0d);
 		editorsPane.setPrefWrapLength(tfPrefWidth * 3 + gap * numInitialFields); //have room for 3 text fields
 		masterPane.minWidth(0d);
-		
+
 		masterPane.getChildren().addAll(btnDecreaseSize, btnIncreaseSize);
 		btnDecreaseSize.setOnAction(new EventHandler<ActionEvent>() {
 			@Override
@@ -63,6 +75,8 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 				editors.add(in);
 				editorsPane.getChildren().add(in);
 				editorsPane.autosize();
+
+				addEditorValueUpdateListener();
 			}
 		});
 		InputField in;
@@ -72,8 +86,16 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 			editorsPane.getChildren().add(in);
 		}
 		editorsPane.autosize();
+
+		addEditorValueUpdateListener();
 	}
-	
+
+	private void addEditorValueUpdateListener() {
+		for (InputField inf : editors) {
+			inf.getValueObserver().addValueListener(editorValueUpdateListener);
+		}
+	}
+
 	@Override
 	public void focusToEditor() {
 		for (InputField tf : editors) {
@@ -84,7 +106,12 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 		}
 		editors.get(0).requestFocus();
 	}
-	
+
+	@Override
+	public ReadOnlyValueObserver<SVStringArray> getReadOnlyObserver() {
+		return valueObserver.getReadOnlyValueObserver();
+	}
+
 	private InputField<ArmaStringChecker, String> getTextField() {
 		InputField<ArmaStringChecker, String> tf = new InputField<>(new ArmaStringChecker());
 		tf.setPrefWidth(tfPrefWidth);
@@ -93,13 +120,18 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 
 	@Override
 	public void submitCurrentData() {
-		for(InputField tf : editors){
+		for (InputField tf : editors) {
 			tf.submitValue();
 		}
 	}
 
 	@Override
 	public SVStringArray getValue() {
+		return valueObserver.getValue();
+	}
+
+	@Nullable
+	private SVStringArray createValue() {
 		String[] values = new String[editors.size()];
 		int i = 0;
 		for (InputField tf : editors) {
@@ -110,20 +142,26 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 		}
 		return new SVStringArray(values);
 	}
-	
+
 	@Override
 	public void setValue(SVStringArray val) {
-		int i = 0;
-		for (String s : val.getAsStringArray()) {
-			editors.get(i++).setValue(s);
+		if (val == null) {
+			for (InputField<ArmaStringChecker, String> editor : editors) {
+				editor.setValue(null);
+			}
+		} else {
+			int i = 0;
+			for (String s : val.getAsStringArray()) {
+				editors.get(i++).setValue(s);
+			}
 		}
 	}
-	
+
 	@Override
 	public @NotNull Node getRootNode() {
 		return masterPane;
 	}
-	
+
 	@Override
 	public void setToOverride(boolean override) {
 		masterPane.getChildren().clear();
@@ -133,7 +171,7 @@ public class ArrayValueEditor implements ValueEditor<SVStringArray> {
 			masterPane.getChildren().add(editorsPane);
 		}
 	}
-	
+
 	@Override
 	public InputField<StringChecker, String> getOverrideTextField() {
 		return overrideField;