+++ src/com/kaylerrenslow/armaDialogCreator/gui/fx/main/controlPropertiesEditor/HexColorValueEditor.java
@@ -13,9 +13,14 @@ package com.kaylerrenslow.armaDialogCreator.gui.fx.main.controlPropertiesEditor;
 import com.kaylerrenslow.armaDialogCreator.control.sv.AHexColor;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.inputfield.InputField;
 import com.kaylerrenslow.armaDialogCreator.gui.fx.control.inputfield.StringChecker;
+import com.kaylerrenslow.armaDialogCreator.util.ReadOnlyValueObserver;
+import com.kaylerrenslow.armaDialogCreator.util.ValueObserver;
+import javafx.event.ActionEvent;
+import javafx.event.EventHandler;
 import javafx.scene.Node;
 import javafx.scene.control.ColorPicker;
 import javafx.scene.layout.StackPane;
+import javafx.scene.paint.Color;
 import org.jetbrains.annotations.NotNull;
 
 /**
@@ -27,6 +32,24 @@ public class HexColorValueEditor implements ValueEditor<AHexColor> {
 	private final InputField<StringChecker, String> overrideField = new InputField<>(new StringChecker());
 	private final StackPane masterPane = new StackPane(colorPicker);
 
+	private final ValueObserver<AHexColor> valueObserver = new ValueObserver<>(null);
+
+	public HexColorValueEditor() {
+		colorPicker.setOnAction(new EventHandler<ActionEvent>() {
+			@Override
+			public void handle(ActionEvent event) {
+				Color newValue = colorPicker.getValue();
+				AHexColor aColor;
+				if (newValue == null) {
+					aColor = null;
+				} else {
+					aColor = new AHexColor(newValue);
+				}
+				valueObserver.updateValue(aColor);
+			}
+		});
+	}
+
 	@Override
 	public void submitCurrentData() {
 
@@ -34,7 +57,7 @@ public class HexColorValueEditor implements ValueEditor<AHexColor> {
 
 	@Override
 	public AHexColor getValue() {
-		return new AHexColor(colorPicker.getValue());
+		return valueObserver.getValue();
 	}
 
 	@Override
@@ -65,5 +88,10 @@ public class HexColorValueEditor implements ValueEditor<AHexColor> {
 	@Override
 	public void focusToEditor() {
 		colorPicker.requestFocus();
+	}
+
+	@Override
+	public ReadOnlyValueObserver<AHexColor> getReadOnlyObserver() {
+		return valueObserver.getReadOnlyValueObserver();
 	}
 }