+++ src/com/kaylerrenslow/armaDialogCreator/util/ReadOnlyValueListener.java
@@ -14,16 +14,15 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
- @author Kayler
- Simple value listener interface.
- Created on 05/31/2016. */
-public interface ValueListener<V> {
+ Created by Kayler on 09/16/2016.
+ */
+public interface ReadOnlyValueListener<V> {
 	/**
-	 The value was updated. However, the value is guaranteed to be different.
-	 
-	 @param observer observer
-	 @param oldValue old value
-	 @param newValue new value
+	The value was updated. However, the value is guaranteed to be different.
+
+	@param observer observer
+	@param oldValue old value
+	@param newValue new value
 	 */
-	void valueUpdated(@NotNull ValueObserver<V> observer, @Nullable V oldValue, @Nullable V newValue);
+	void valueUpdated(@NotNull ReadOnlyValueObserver<V> observer, @Nullable V oldValue, @Nullable V newValue);
 }